refactor: Rename project to "PopMap Recommender" and update AboutPage with new content and styles
